Fix ament_cpplint [runtime/int] errors in coverage_executor_node
Coverage-executor callbacks used "unsigned long long" + "%llu" to print a
std::uint64_t child_gen_ token.  ament_cpplint [runtime/int] rejects
C-style "long" / "long long" types, so this triggered 4 lint failures
(cpp:711/712/778/779) and broke CI on the main branch after the showcase
commit.

- <cinttypes> added for PRIu64
- format spec switched to %" PRIu64 "
- redundant static_cast<unsigned long long>(...) removed; values are
  already std::uint64_t

Verified locally: ament_cpplint returns rc 0 (No problems found) for
tunnel_coverage_executor; tunnel_map_core, tunnel_coverage_msgs (N/A no
C++), tunnel_coverage_planner also clean.  The xmllint timeout seen in
local colcon test is an offline-network sandbox issue (download.ros.org
XSD unreachable from WSL); CI container has network access and xmllint
passed before this fix.

No tag or red-line branch touched.  Will fast-forward main and push
once CI is green.

+++ tunnel_coverage_executor/src/coverage_executor_node.cpp
@@ -17,6 +17,7 @@
 #include <tf2/exceptions.h>
 
 #include <algorithm>
+#include <cinttypes>
 #include <cmath>
 #include <filesystem>
 #include <memory>
@@ -707,9 +708,9 @@ void CoverageExecutorNode::sendNavigate(
     {
       if (gen != child_gen_) {
         RCLCPP_WARN(get_logger(),
-          "Stale NavigateToPose result ignored (gen %llu != %llu)",
-          static_cast<unsigned long long>(gen),
-          static_cast<unsigned long long>(child_gen_));
+          "Stale NavigateToPose result ignored (gen %" PRIu64 " != %" PRIu64 ")",
+          gen,
+          child_gen_);
         return;
       }
       const bool ok = r.code == rclcpp_action::ResultCode::SUCCEEDED;
@@ -774,9 +775,9 @@ void CoverageExecutorNode::sendFollow(
     {
       if (gen != child_gen_) {
         RCLCPP_WARN(get_logger(),
-          "Stale FollowPath result ignored (gen %llu != %llu)",
-          static_cast<unsigned long long>(gen),
-          static_cast<unsigned long long>(child_gen_));
+          "Stale FollowPath result ignored (gen %" PRIu64 " != %" PRIu64 ")",
+          gen,
+          child_gen_);
         return;
       }
       const bool ok = r.code == rclcpp_action::ResultCode::SUCCEEDED;
